rewrite/convert mod over to param2 colorization New dependency:  Unified Dyes
Neutral-to-positive side effects:

Crafting has changed: only white dye is used to create a patterned
node.  Once created, this node can be placed and then painted
normally with right-click.

Now all of Unified Dyes' colors can be applied

A few colors were re-mapped in the conversion LBM to keep the colors
as similar as possible to Unified Dyes' palette.